api/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pdf-image-size-detector-api"
version = "1.0.0"
description = "API for extracting images from PDFs and detecting their sizes"
dependencies = [
    "flask==2.3.3",
    "flask-cors==4.0.0",
    "Pillow==10.0.1",
    "PyPDF2==3.0.1",
    "pdf2image==1.16.3",
    "python-dotenv==1.0.0",
    "gunicorn==21.2.0"
]

[tool.hatch.build.targets.wheel]
packages = ["api"]
